Lab 1
Create a program that provides a menu with 5 options to the user (4 operations and a “Quit” option). The 4 operations that you need to implement are the following:
￼
Asks the user to enter the 3 dimensions of a box and displays back to the user: Volume and Surface Area

Even/Odd: Given a positive integer number as an input. Tell the user whether the number is even or odd, and display all the even or odd numbers from 0 to that number (depending on the input).

Factorial: get a positive number from the user then the factorial(n!) of a number is displayed. n! is calculated as:
n*(n-1)*(n-2) .... (2)(1)
Ex: 4! = 4*3*2*1 = 24

X^n means that you are going to multiply X n times. Ex: 2^3 = 2*2*2 = 8, 3^4 = 3*3*3*3= 81
Ask the user for two values (X and n) and raise X to the nth power.

Quit:Terminates the program execution (this means that your program should have a menu that repeats until the option 5 is selected.